# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Backend Funeraria App API
  version: 2.1.0
  description: |
    Contrato HTTP del monolito `backend-funeraria-app`.

    La API esta organizada por slices funcionales y documenta:
    - endpoints disponibles hoy en controllers reales
    - requests y responses con sus DTOs actuales
    - reglas de seguridad por JWT ligado a dispositivo
    - errores esperables por validacion, autorizacion y reglas de negocio

    Convenciones importantes:
    - Los endpoints autenticados requieren `Authorization: Bearer <token>`.
    - Los endpoints autenticados tambien requieren `X-Device-Id`.
    - `POST /api/v1/users/login` y `POST /api/v1/users/refresh` aceptan `Idempotency-Key`.
    - Todas las respuestas incluyen `X-Trace-Id`; si el cliente envia `X-Correlation-Id`, la API lo propaga.
    - Los errores de negocio y validacion siguen el modelo `ProblemDetail`.
servers:
  - url: http://localhost:8081
    description: Entorno local
...
paths:
...
  /api/v1/users/rol/{email}:
    post:
      tags: [Users]
      summary: Agregar o actualizar rol de un usuario
      description: Solo ADMIN. Devuelve el set de roles actualizado del usuario.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/EmailPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolRequestDto'
      responses:
        '200':
          description: Roles actualizados.
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/RolRequestDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /api/v1/items:
    get:
      tags: [Items]
      summary: Listar items
      description: Disponible para ADMIN y USER.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
      responses:
        '200':
          description: Lista de items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    post:
      tags: [Items]
      summary: Crear item
      description: Solo ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequestDto'
      responses:
        '201':
          description: Item creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponseDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /api/v1/items/{code}:
    get:
      tags: [Items]
      summary: Obtener item por codigo
      description: Solo ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/ItemCodePath'
      responses:
        '200':
          description: Item encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    put:
      tags: [Items]
      summary: Actualizar item
      description: Solo ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/ItemCodePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequestDto'
      responses:
        '200':
          description: Item actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponseDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    delete:
      tags: [Items]
      summary: Eliminar item
      description: Solo ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/ItemCodePath'
      responses:
        '200':
          description: Operacion completada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatusModel'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /api/v1/items/category/{id}:
    get:
      tags: [Items]
      summary: Listar items por categoria
      description: Disponible para ADMIN y USER.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/LongIdPath'
      responses:
        '200':
          description: Items de la categoria indicada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /api/v1/items/{code}/image/upload:
    post:
      tags: [Items]
      summary: Cargar imagen de item
      description: Solo ADMIN. Sube la imagen al proveedor de almacenamiento configurado y actualiza el item.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/ItemCodePath'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen cargada correctamente.
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de acceso emitido por `/api/v1/users/login`.
  parameters:
    XDeviceId:
      name: X-Device-Id
      in: header
      required: true
      description: Identificador estable del dispositivo o cliente que creo la sesion.
      schema:
        type: string
        example: web-chrome-gabriel
...
    LongIdPath:
      name: id
      in: path
      required: true
      description: Identificador numerico.
      schema:
        type: integer
        format: int64
        example: 1
    EmailPath:
      name: email
      in: path
      required: true
      description: Email del usuario.
      schema:
        type: string
        format: email
        example: admin@funeraria.local
...
    ItemCodePath:
      name: code
      in: path
      required: true
      description: Codigo del item.
      schema:
        type: string
        example: ATAUD-ROBLE-01
...
  responses:
    ValidationResponse:
      description: Error de validacion de request.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
          example:
            type: about:blank
            title: Validacion fallida
            status: 400
            detail: La solicitud contiene datos invalidos.
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
            errors:
              - field: email
                message: El email es requerido.
    UnauthorizedResponse:
      description: Request sin credenciales validas, token invalido o dispositivo no coincidente.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Error de aplicacion
            status: 401
            detail: El token no coincide con el dispositivo actual.
            code: security.jwt.device.mismatch
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
            correlationId: corr-request-0001
    ForbiddenResponse:
      description: El usuario autenticado no tiene permisos para ejecutar la operacion.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Acceso denegado
            status: 403
            detail: Access Denied
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    NotFoundResponse:
      description: Recurso no encontrado.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Error de aplicacion
            status: 404
            detail: No existe afiliado con el DNI especificado.
            code: affiliate.error.not.found
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    ConflictResponse:
      description: Conflicto de negocio, duplicado o estado incompatible.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Error de aplicacion
            status: 409
            detail: Ya existe un afiliado registrado con el DNI especificado.
            code: affiliate.error.dni.already.exists
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
...
    InternalServerErrorResponse:
      description: Error interno del servidor.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Error interno del servicio
            status: 500
            detail: Ocurrio un error interno inesperado.
            code: error.internal
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
  headers:
    TraceIdHeader:
      description: Identificador unico de la request generado o propagado por el servicio.
      schema:
        type: string
        example: 4bf92f3577b34da6a3ce929d0e0e4736
    CorrelationIdHeader:
      description: Identificador opcional provisto por el cliente para correlacionar requests entre sistemas.
      schema:
        type: string
        example: corr-request-0001
  schemas:
    ProblemDetail:
      type: object
      description: Modelo de error estandar usado por el backend.
      example:
        type: about:blank
        title: Error de aplicacion
        status: 404
        detail: No existe afiliado con el DNI especificado.
        code: affiliate.error.not.found
        traceId: 4bf92f3577b34da6a3ce929d0e0e4736
        correlationId: corr-request-0001
      properties:
        type:
          type: string
          format: uri
          nullable: true
          example: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
          nullable: true
        code:
          type: string
          description: Clave funcional vinculada a `messages_es.properties`.
        traceId:
          type: string
          description: Identificador unico de la request para correlacionar logs, errores y metricas.
          example: 4bf92f3577b34da6a3ce929d0e0e4736
        correlationId:
          type: string
          nullable: true
          description: Identificador opcional provisto por el cliente y propagado por el backend.
          example: corr-request-0001
        retryAfterSeconds:
          type: integer
          nullable: true
      additionalProperties: false
    ValidationProblem:
      example:
        type: about:blank
        title: Validacion fallida
        status: 400
        detail: La solicitud contiene datos invalidos.
        traceId: 4bf92f3577b34da6a3ce929d0e0e4736
        errors:
          - field: email
            message: El email es requerido.
      allOf:
        - $ref: '#/components/schemas/ProblemDetail'
        - type: object
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: no debe estar vacio
      required: [field, message]
...
    OperationStatusModel:
      type: object
      properties:
        name:
          type: string
          example: DELETE ITEM
        result:
          type: string
          example: SUCCESSFUL
      required: [name, result]
...
    BrandDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        webPage:
          type: string
          example: https://brand.example
...
    BrandResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        webPage:
          type: string
    CategoryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
...
    CategoryResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
...
    RolRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          enum: [ROLE_ADMIN, ROLE_USER]
      required: [name]
...
    ItemRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        code:
          type: string
        price:
          type: number
          example: 150000.00
        itemLength:
          type: number
        itemHeight:
          type: number
        itemWidth:
          type: number
        brand:
          $ref: '#/components/schemas/BrandDto'
        category:
          $ref: '#/components/schemas/CategoryDto'
      required: [name, price]
    ItemResponseDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        code:
          type: string
        itemImageLink:
          type: string
          example: https://storage.example/items/ataud-roble-01.png
        stock:
          type: integer
        price:
          type: number
        itemLength:
          type: number
        itemHeight:
          type: number
        itemWidth:
          type: number
        category:
          $ref: '#/components/schemas/CategoryResponseDto'
        brand:
          $ref: '#/components/schemas/BrandResponseDto'
...